+++ CHANGELOG.md
@@ -6,7 +6,9 @@ All notable changes to this project will be documented in this file. This projec
 
 ### Fixed
 - [#138](https://github.com/cloudcreativity/laravel-json-api/issues/138)
-Generator commands not working since Laravel v5.5.28
+Generator commands not working since Laravel v5.5.28.
+- [#131](https://github.com/cloudcreativity/laravel-json-api/issues/131)
+Ensure Eloquent adapter uses default pagination parameters.
 
 ## [0.11.3] - 2017-12-01